+++ tool/restructure.js
@@ -13,13 +13,13 @@ make sure you update the dataFile location and name
 
 ************************************************************************/
 
-var dataFile = fs.readFileSync(process.cwd() + '/backup-1486398812650.json');
+var dataFile = fs.readFileSync(process.cwd() + '/backup-1486435528386.json');
+
 var data = JSON.parse(dataFile);
+
 var newData = {
   triggers : {}
 };
-
-var keys = Object.keys(data);
 
 newData.users = data.dev.users;
 newData.lastTrigger = data.lastTrigger; 
@@ -29,7 +29,8 @@ newData.song_stats = data.song_stats;
 
 Object.keys(data.triggers).forEach((key)=>{
   var thisTrig = data.triggers[key];
-  newData.triggers[thisTrig.Trigger.toLowerCase()] = thisTrig;
+  thisTrig.Trigger = thisTrig.Trigger.toLowerCase();
+  newData.triggers[key] = thisTrig;
 });
 
 fs.writeFileSync(process.cwd() + '/restructured-data.json', JSON.stringify( newData ));